acpi: add ACPI support in Firecracker
Initial ACPI setup for x86.

We have FADT in Hardware Reduced mode. This means that we declare to the
guest that we don't implement in "hardware" some features, such as
control and event registers.

We also have MADT, which essentially replaces the functionality of
MPTable; we define on IO-APIC and one LocalAPIC structure per vCPU. All
vCPUs are marked as online by default.

Finally, we have an empty DSDT. This will be populated with info about
VirtIO, legacy and ACPI devices in future commits.

Signed-off-by: Babis Chalios <bchalios@amazon.es>