acpi: fix test_net_change_mac_address test
This test is looking for the MMIO address range of virtio-net devices
and then uses it to update the MAC address of the devices by writing
under `/dev/mem`.

The test was using the kernel command line to find the address range.
With ACPI we don't pass this range via kernel command line any more, so
adapt the test to look for this information under /proc/iomem.

Signed-off-by: Babis Chalios <bchalios@amazon.es>