pci: add support for PCIe segment
Add a PCIe segment which includes a single PCIe root port and a bus.
At the moment, we always enabled it.

At the moment, the PCIe segment is always enabled. Later commit will
make it optional and enable it only when a command line argument flag is
passed to Firecracker binary.

Signed-off-by: Babis Chalios <bchalios@amazon.es>