vmclock: add snapshot safety features
Add support for `vm_generation_counter` field in VMClock ABI. This field
is similar to `disruption_marker` but it's only updated on snapshot
loading events (not in live migration). It is meant to provide the guest
with snapshot safety notifications.

Moreover, add support for the notification capability. This capability
require us to send an ACPI notification every time we change the
seq_count field to a new even value. This essentially means that we need
to send a notification upon resuming from a snapshot just before
resuming vCPUs.

Signed-off-by: Babis Chalios <bchalios@amazon.es>